Refactor Meeting class into package module
